feat(algorithms): move queue into common
