+++ crates/algorithms/src/shortest_paths/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod queue;
